+++ crates/vela_reflect/src/member_records.rs
@@ -24,7 +24,7 @@ pub(crate) fn method_record_with_owner(type_name: &str, method: &MethodDesc) -> 
 
 fn method_record_fields(method: &MethodDesc) -> ReflectFields {
     let mut fields = BTreeMap::new();
-    fields.insert("id".to_owned(), id_value(method.id.get()));
+    fields.insert("id".to_owned(), id_value(u128::from(method.id.get())));
     fields.insert("name".to_owned(), string(method.name.clone()));
     fields.insert("origin".to_owned(), origin_value(method.origin));
     fields.insert(
@@ -209,11 +209,10 @@ fn field_record_fields(field: &FieldDesc) -> ReflectFields {
     fields
 }
 
-fn id_value(id: u64) -> ReflectValue {
-    // TODO(reflect): stable IDs are u64, but reflection currently exposes IDs
+fn id_value(id: u128) -> ReflectValue {
+    // TODO(reflect): DefIds are u128, but reflection currently exposes IDs
     // through signed script ints. Replace this lossy saturation with a deliberate
-    // unsigned/ID value surface before treating reflect::id() as a stable public
-    // identity API.
+    // ID value surface before treating reflect::id() as a stable public identity API.
     int_value(i64::try_from(id).unwrap_or(i64::MAX))
 }
